Build:WIP:IImproved loaders with the abandonment of polib for the babel library. Systematization of the use of str in the upper layers of loaders and management in the lower layers of the use of Path() or str. Preparation for use of high-level functions (interface loaded in __init__.py).
