+++ src/i18n_tools/loaders/loader.py
@@ -21,7 +21,6 @@ modifying this file, you agree to abide by the terms of this license.
 This module is authored and maintained as part of the i18n-tools package.
 """
 
-import tarfile
 from pathlib import Path
 from typing import Any, Dict, List
 
@@ -32,7 +31,6 @@ from babel.messages.pofile import read_po, write_po
 from i18n_tools.loaders.utils import (
     _create_gzip,
     _load_json,
-    _non_traversal_path,
     _save_json,
 )
 
@@ -322,81 +320,3 @@ def save_locale_pot(file_path: str, po_data: Catalog) -> None:
     :return:
     """
     _save_pot(file_path, po_data)
-
-
-def create_module_archive(root_path: str, module: str, archive_name: str) -> None:
-    """
-    Create a tar.gz archive of a module's contents, including all subdirectories and files,
-    by identifying the top-level module from the given module/package path.
-
-    :param root_path: The base path where the modules are located.
-    :type root_path: str
-    :param module: The module or module/package path (e.g., "mod-1/pkg-1").
-    :type module: str
-    :param archive_name: The name of the archive file.
-    :type archive_name: str
-    :return: None
-    :rtype: None
-    """
-
-    # Convert root_path to a Path object and resolve it to an absolute path
-    # This ensures that the path is absolute and any symbolic links are resolved
-    root_path = Path(root_path).resolve()
-
-    # Identify the top-level module from the given module path
-    # This is done by splitting the module path and taking the first part
-    # For example, if module is "mod-1/pkg-1", top_level_module will be "mod-1"
-    top_level_module = module.split("/")[0]
-
-    # Construct the full path to the top-level module
-    # This is the directory that will be archived
-    module_path = (root_path / top_level_module).resolve()
-
-    # Construct the path for the archive file
-    # The archive will be created in the root_path directory with the given archive_name
-    archive_path = root_path / f"{archive_name}.tar.gz"
-
-    # Open the archive file in write mode with gzip compression
-    with tarfile.open(archive_path, "w:gz") as tar:
-        # Add the top-level module to the archive
-        # The arcname parameter ensures that the directory structure is preserved in the archive
-        # This means the archive will contain the top-level module directory and all its contents
-        tar.add(module_path, arcname=top_level_module)
-
-
-def restore_module_from_archive(root_path: str, module: str, archive_name: str) -> None:
-    """
-    Restore a module's contents from a tar.gz archive.
-
-    :param root_path: The base path where the modules are located.
-    :type root_path: str
-    :param module: The module or sub-module path (e.g., "mod-1/pkg-1").
-    :type module: str
-    :param archive_name: The name of the archive file.
-    :type archive_name: str
-    """
-
-    # Split the module path to get the top-level module
-    module_parts = module.split("/")
-    top_level_module = module_parts[0]
-
-    # Construct the path for the archive file
-    archive_path = Path(root_path) / f"{archive_name}.tar.gz"
-
-    # Construct the full path to the module
-    module_path = Path(root_path) / module
-
-    # Check if the archive file exists
-    if archive_path.exists():
-        # Open the archive file in read mode with gzip compression
-        with tarfile.open(archive_path, "r:gz") as tar:
-            # Extract all members that are safe to extract
-            tar.extractall(
-                path=root_path,
-                members=_non_traversal_path(
-                    root_path, [top_level_module], tar.getmembers()
-                ),
-            )
-    else:
-        # Raise an error if the archive file does not exist
-        raise FileNotFoundError(f"Archive file '{archive_path}' not found.")
